pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dna-kyc-screener"
version = "0.1.0"
description = "Rule-based KYC screening for DNA synthesis orders (research/portfolio)"
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.83",
    "rapidfuzz>=3.6",
    "fastapi>=0.110",
    "pydantic>=2.6",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "uvicorn[standard]>=0.29",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["api", "audit", "core"]
